--- src/main/java/org/songeun/petdongne_server/addess/infrastructure/elasticsearch/document/converter/AddressHierarchyToIntegerConverter.java
@@ -1,15 +1,0 @@
-package org.songeun.petdongne_server.addess.infrastructure.elasticsearch.document.converter;
-
-import org.songeun.petdongne_server.addess.infrastructure.elasticsearch.document.AddressHierarchy;
-import org.springframework.core.convert.converter.Converter;
-import org.springframework.data.convert.WritingConverter;
-
-@WritingConverter
-public class AddressHierarchyToIntegerConverter implements Converter<AddressHierarchy, Integer> {
-
-    @Override
-    public Integer convert(AddressHierarchy hierarchy) {
-        return hierarchy.getLevel();
-    }
-
-}
