--- src/main/java/org/songeun/petdongne_server/addess/infrastructure/elasticsearch/document/converter/AddressTypeToStringConverter.java
@@ -1,16 +1,0 @@
-package org.songeun.petdongne_server.addess.infrastructure.elasticsearch.document.converter;
-
-import org.songeun.petdongne_server.addess.infrastructure.elasticsearch.document.AddressType;
-import org.springframework.core.convert.converter.Converter;
-import org.springframework.data.convert.WritingConverter;
-
-@WritingConverter
-public class AddressTypeToStringConverter implements Converter<AddressType, String> {
-
-    @Override
-    public String convert(AddressType type) {
-        return type.getKoreanName();
-
-    }
-
-}
