--- src/main/java/org/songeun/petdongne_server/addess/infrastructure/elasticsearch/document/repository/ElasticsearchAddressRepository.java
@@ -1,10 +1,0 @@
-package org.songeun.petdongne_server.addess.infrastructure.elasticsearch.document.repository;
-
-import org.songeun.petdongne_server.addess.infrastructure.elasticsearch.document.AddressDocument;
-import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ElasticsearchAddressRepository extends ElasticsearchRepository<AddressDocument, String> {
-
-}
