--- src/main/java/org/songeun/petdongne_server/addess/infrastructure/elasticsearch/index/AddressIndexRepository.java
@@ -1,9 +1,0 @@
-package org.songeun.petdongne_server.addess.infrastructure.elasticsearch.index;
-
-public interface AddressIndexRepository {
-
-    public boolean existIndex();
-
-    public boolean createIndex();
-
-}
