+++ src/main/java/org/songeun/petdongne_server/address/infrastructure/elasticsearch/index/AddressIndexRepository.java
@@ -1,0 +1,9 @@
+package org.songeun.petdongne_server.address.infrastructure.elasticsearch.index;
+
+public interface AddressIndexRepository {
+
+    public boolean existIndex();
+
+    public boolean createIndex();
+
+}
