+++ src/test/java/org/songeun/petdongne_server/address/infrastructure/elasticsearch/document/AddressTypeTest.java
@@ -2,11 +2,10 @@ package org.songeun.petdongne_server.address.infrastructure.elasticsearch.docume
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.songeun.petdongne_server.addess.infrastructure.elasticsearch.document.AddressType;
 import org.songeun.petdongne_server.global.exception.BusinessException;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.songeun.petdongne_server.addess.infrastructure.elasticsearch.error.AddressErrorStatus.ADDRESS_TYPE_NOT_FOUND;
+import static org.songeun.petdongne_server.address.infrastructure.elasticsearch.error.AddressErrorStatus.ADDRESS_TYPE_NOT_FOUND;
 
 class AddressTypeTest {
